Refactor backend routes into separate modules
