+++ backend/routes/main.py
@@ -1,0 +1,8 @@
+from flask import Blueprint
+
+bp = Blueprint('main', __name__)
+
+
+@bp.route('/')
+def index():
+    return {"message": "Hello World"}
